GH-110109: Move tests for pathlib ABCs to new module.
These tests will be used in a standalone PyPI package.
